Narrowly define what a positive window is
Previously as long as the CG of the player was within the window, that window
would be labeled positive. Now we define a method with a stricter criterion -
the CG has to be close to the center of the window for the window to be
considered positive. This reduces the number of positive examples that barely
contain the character, thereby reducing false positives.